backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI & LLM - Google Gemini
google-generativeai==0.3.2
chromadb==0.4.22

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Caching & Queue
redis==5.0.1
celery==5.3.6

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# CORS
python-cors==1.0.0

# Date & Time
python-dateutil==2.8.2

# Validation
email-validator==2.1.0
phonenumbers==8.13.27
